+++ Peck/PADropdownViewController.m
@@ -111,7 +111,7 @@
 
 - (void) barDidSelectItemAtIndex:(NSInteger)index
 {
-    NSLog(@"%d", index);
+    NSLog(@"%ld", (long)index);
     UIViewController * oldVC = self.activeViewController;
     UIViewController * newVC = self.secondaryViewControllers[index];
 
@@ -149,7 +149,7 @@
 
 - (void) barDidDeselectItemAtIndex:(NSInteger)index
 {
-    NSLog(@"%d", index);
+    NSLog(@"%ld", (long)index);
     UIViewController * oldVC = self.activeViewController;
     UIViewController * newVC = self.primaryViewController;
 
@@ -187,7 +187,7 @@
 
 - (void) barDidSlideLeftToIndex:(NSInteger)index
 {
-    NSLog(@"%d", index);
+    NSLog(@"%ld", (long)index);
     UIViewController * oldVC = self.activeViewController;
     UIViewController * newVC = self.secondaryViewControllers[index];
 
@@ -227,7 +227,7 @@
 
 - (void) barDidSlideRightToIndex:(NSInteger)index
 {
-    NSLog(@"%d", index);
+    NSLog(@"%ld", (long)index);
     UIViewController * oldVC = self.activeViewController;
     UIViewController * newVC = self.secondaryViewControllers[index];
